Feature: Movable objects moving and render

+++ src/main/java/aircraftgame/Aircraft.java
@@ -3,7 +3,7 @@ package aircraftgame;
 public abstract class Aircraft extends MovableObject {
 	int health;
 	
-	Aircraft(int xPosition, int yPosition, float xDirection, float yDirection){
+	Aircraft(float xPosition, float yPosition, float xDirection, float yDirection){
 		super(xPosition, yPosition, xDirection, yDirection);
 		health = 100;
 	}
